# docs/swagger.yaml
basePath: /
definitions:
  internal_app.App:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  internal_app.CreateAppRequest:
    properties:
      name:
        type: string
      status:
        type: integer
    required:
    - name
    type: object
  internal_app.UpdateAppRequest:
    properties:
      name:
        type: string
      status:
        type: integer
    type: object
  internal_user.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_user.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/internal_user.User'
    type: object
  internal_user.CreateUserRequest:
    properties:
      app_id:
        type: integer
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - app_id
    - email
    - firstname
    - lastname
    - password
    type: object
  internal_user.UpdateUserRequest:
    properties:
      app_id:
        type: integer
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 8
        type: string
      status:
        type: integer
    type: object
  internal_user.User:
    properties:
      app_id:
        type: integer
      app_name:
        type: string
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      password:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  keeper_pkg_render.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a microservice for user management.
  title: Keeper API
  version: "1.0"
paths:
  /apps:
    get:
      description: Get a list of all registered apps
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_app.App'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: List all apps
      tags:
      - apps
    post:
      consumes:
      - application/json
      description: Create a new app with the provided details
      parameters:
      - description: App details
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/internal_app.CreateAppRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_app.App'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Create a new app
      tags:
      - apps
  /apps/{id}:
    delete:
      description: Remove an app from the system by ID
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Delete app
      tags:
      - apps
    get:
      description: Get a single app by its unique ID
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_app.App'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Get app by ID
      tags:
      - apps
    put:
      consumes:
      - application/json
      description: Update an existing app's details
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated app details
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/internal_app.UpdateAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_app.App'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Update app
      tags:
      - apps
  /health:
    get:
      description: Get the health status of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      summary: Check service health
      tags:
      - health
  /users:
    get:
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_user.User'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Remove a user from the system by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a single user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user's details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - users
  /users/auth:
    post:
      consumes:
      - application/json
      description: Login with email and password to receive a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_user.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keeper_pkg_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_user.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keeper_pkg_render.Response'
      summary: Authenticate user
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"